Added my own capture programs.
Having problems getting frames into python. Created the capture.cpp, a
cut-down version of blob.cpp which just captures from the camera and
dumps to disk, but scanner.c is written as a python object, so it'd be
much easier to use it in python, but I'm unable to get any images into
python.