Working at about 1Hz.
caponeshot grabs webcam images and dumps to a file.
slowtest loads the file, runs Tridge's algoritm on it, highlights
regions of interest and dumps the file to webroot. Not much faster by
omitting highlighting and disk access. I suspect the slowness is in the
memory alocation after resizing the image. Not sure how smart python is.